% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggg.R
\name{gg}
\alias{gg}
\title{Pass ggplotall to plotly (basic)}
\usage{
gg(a = (ggplot2::last_plot() + theme_bw()), b = c("all"))
}
\arguments{
\item{a}{item to pass}

\item{b}{fields to pass}
}
\value{
returns ggplotly viewer object
}
\description{
Pass ggplotall to plotly (basic)
}
\examples{
print("test")
}
