% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggg.R
\name{ggg}
\alias{ggg}
\title{Pass ggplotall to plotly (trimmed)}
\usage{
ggg(a = (ggplot2::last_plot() + theme_bw()), b = c("text"))
}
\arguments{
\item{a}{item to pass}

\item{b}{field to pass}
}
\value{
returns ggplotly viewer object
}
\description{
Pass ggplotall to plotly (trimmed)
}
\examples{
print("test")
}
