% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplotall.R
\name{ggplotall}
\alias{ggplotall}
\title{Main plotter function}
\usage{
ggplotall(
  df,
  xval,
  yval,
  Pop = allidv_gen@pop,
  xl = xval,
  yl = yval,
  title = "",
  labs = PopLabs,
  colrs = ColLabs,
  text = "",
  force = 3,
  force_pull = 1,
  max.overlaps = 200
)
}
\arguments{
\item{df}{dataframe (tibble) to generate plot from}

\item{xval}{name of x data column}

\item{yval}{name of y data column}

\item{Pop}{vector of pops drawn from}

\item{xl}{x axis label (text)}

\item{yl}{y axis label (text)}

\item{title}{title for graph}

\item{labs}{text labels for pops}

\item{colrs}{colors for pops}

\item{text}{text for plotly}

\item{force}{internal value for repel}

\item{force_pull}{internal value for repel}

\item{max.overlaps}{internal value for repel}
}
\value{
returns ggplot graph (or paints it)
}
\description{
Main plotter function
}
\examples{
print("test")
}
