% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sum.R
\name{label_sum}
\alias{label_sum}
\title{Label sum}
\usage{
label_sum(Pop, xvals, yvals, text)
}
\arguments{
\item{Pop}{vector}

\item{xvals}{vector}

\item{yvals}{vector}

\item{text}{vector}
}
\value{
tibble of processed values for use in graphing
}
\description{
Label sum
}
\examples{
print("none")
}
