% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_template_ref.R
\name{load_data_ref}
\alias{load_data_ref}
\title{Title}
\usage{
load_data_ref(filename, template)
}
\arguments{
\item{filename}{path to reference file}

\item{template}{object of class TempRef}
}
\value{
returns object of class DataRef
}
\description{
Title
}
\examples{
print("test")
}
