% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_template_ref.R
\name{load_template_ref}
\alias{load_template_ref}
\title{Generate master template from file...}
\usage{
load_template_ref(filename, ...)
}
\arguments{
\item{filename}{path to reference file}

\item{...}{vector of labels to target}
}
\value{
returns template object for future functions
}
\description{
Generate master template from file...
}
\examples{
print("test")
}
