% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_template_ref.R
\name{make_template_ref}
\alias{make_template_ref}
\title{Convert ref frame to ref template}
\usage{
make_template_ref(reftab)
}
\arguments{
\item{reftab}{reference tibble}
}
\value{
returns template object for future functions
}
\description{
Convert ref frame to ref template
}
\examples{
print("test")
}
