% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_interpolator_pop.R
\name{na_interpolator_pop}
\alias{na_interpolator_pop}
\title{Interpolates genind by pop then spp. and converts to matrix}
\usage{
na_interpolator_pop(genind)
}
\arguments{
\item{genind}{genind object}
}
\value{
returns matrix
}
\description{
Interpolates genind by pop then spp. and converts to matrix
}
\examples{
print("test")
}
