% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_monos.R
\name{no_monos}
\alias{no_monos}
\title{Process dataframe from upstream processing}
\usage{
no_monos(rawdata, na_val = "000000")
}
\arguments{
\item{rawdata}{dataframe input}

\item{na_val}{value coding NA data in dataframe}
}
\value{
returns filtered dataset (removing monomorphic sites)
}
\description{
Process dataframe from upstream processing
}
\examples{
print("test")
}
