% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_genraw}
\alias{read_genraw}
\title{Read raw gendata (from mj pipeline)}
\usage{
read_genraw(filename)
}
\arguments{
\item{filename}{path to input raw gendata file}
}
\value{
returns tibble of nomono-filtered data, sorted by Indv (matches targetid)
}
\description{
Read raw gendata (from mj pipeline)
}
\examples{
print("test")
}
