% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_genref}
\alias{read_genref}
\title{Read genref (from mj pipeline)
Still missing regionr & relevlr steps... include later}
\usage{
read_genref(filename)
}
\arguments{
\item{filename}{path to impot genref file}
}
\value{
returns reference tibble (matched to idv)
}
\description{
Read genref (from mj pipeline)
Still missing regionr & relevlr steps... include later
}
\examples{
print("test")
}
