% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_sum.R
\name{repel_all}
\alias{repel_all}
\title{Helper function for the main game (ggplotall)}
\usage{
repel_all(
  xvals,
  yvals,
  Pop = allidv_gen@pop,
  levs = PopLevs,
  text = "",
  force = 3,
  force_pull = 1,
  max.overlaps = 200
)
}
\arguments{
\item{xvals}{x axis values}

\item{yvals}{y axis values}

\item{Pop}{pop list}

\item{levs}{pop levs}

\item{text}{text for plotly functions}

\item{force}{for geom_repel}

\item{force_pull}{for geom_repel}

\item{max.overlaps}{for geom_repel}
}
\value{
returns repel object
}
\description{
Helper function for the main game (ggplotall)
}
\examples{
print("none")
}
