% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler}
\alias{scaler}
\title{Van specific scaler}
\usage{
scaler(ref_frame)
}
\arguments{
\item{ref_frame}{reference tibble for graph}
}
\value{
character vector for text to plotly
}
\description{
Van specific scaler
}
\examples{
print("test")
}
