% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler2}
\alias{scaler2}
\title{Generic scaler}
\usage{
scaler2(ref_frame)
}
\arguments{
\item{ref_frame}{reference tibble for graph}
}
\value{
character vector for text to plotly
}
\description{
Generic scaler
}
\examples{
print("test")
}
