% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_popcrunchr.R
\name{theme_popcrunchr}
\alias{theme_popcrunchr}
\title{popcrunchr theme}
\usage{
theme_popcrunchr(vals = ColLabs)
}
\arguments{
\item{vals}{list of values for fill scale}
}
\value{
returns scale_fill_manual list for use with ggplot
}
\description{
popcrunchr theme
}
\examples{
theme_popcrunchr()
}
